renames the blueprints and tests
