+++ api/books/__init__.py
@@ -1,6 +1,6 @@
 from flask import Blueprint
 
-book = Blueprint('book', __name__)
+book = Blueprint('books', __name__)
 
 from . import views
